Ensure we do not log more than the maximum API gateway can handle (#7)
Since the logger writes batches of records to the API, we need to ensure
we do not send more than the 10 MB limit of API gateway. This change
simply predicts the batch length by serializing JSON values before
sending them off.

There is a more efficient version of this change that does a streaming
serialization of JSON values (and concatenates them together with `,`),
but I think we can explore that later if needed.