Fix race conditions in logger.
Uncovered by unittests in the main braintrust repo.

It is a requirement that rows logged through the SDK in a single thread must be
flushed to the lambda function in exactly that order, to ensure the DB can keep
track of the ordering. However, the python SDK logger has a background thread to
flush logs, which could race with the `flush` method (invoked by
`experiment.summarize()`), leading to a violation of this requirement. A few
issues were uncovered:

The `_publisher` thread was popping an item off the queue before flushing. If
`flush` is invoked concurrently, this could lose the ordering of the initial
item with respect to the rest of the queue. We address this by not popping
anything off the queue in `_publisher` and just directly invoking `flush` in a
loop.

With the above fix, now the `_publisher` loop was spending a lot of time
spinning (or publishing very small batches of records). To address this, we
introduce a `threading.Event`, waited on by the `_publisher` thread and signaled
by the `log` method whenever the queue size reaches the default batch size.