JS SDK: Launch logging requests concurrently. (#15)
As long as the row IDs in a set of rows are independent, we can batch
them up and log them concurrently. The only extra step we need then is
to merge rows with the same ID before logging, which saves on bandwidth
anyways.

Aside from sanity-check testing, verified on prod that the
`stress-tests/large_experiment.ts` benchmark, run with 1 experiment of
5000 rows shows improvement. Before, it took about 321.26 seconds. After
it takes about 99.77 seconds.