Rework the codegen.
Instead of generating the full file, we are going to try generating just
the "interface" of the SDK, i.e. the function/class signatures,
documentation, and type definitions.

We don't generate a true "interface" file, because the implementation
may want to manage state and methods differently. Instead, the public
file calls into internal definitions in a "logger_impl" file.

We put out a small example, starting to implement the "Dataset" class in
`codegen/example-output`.