Fix deadlock in python logger.
The condition variable approach was flawed because
`condition_variable.wait()` will only respond to notify signals sent
after the `wait` runs. So if the logger thread was in the middle of a
flush while the producer thread exhausts the queue and sends a notify
signal, the consumer thread will miss the notify signal on its next
`wait` call and we have a deadlock.

The solution is to replace the condition variable with a semaphore. So
that the producer can increment the semaphore as a "persistent" way to
trigger the consumer thread, which blocks on the semaphore being 0.