A couple testing-related improvements (#40)
- The `braintrust eval` python command can emulate the `python -m
unittest` command by adding the directory python was invoked in to
`sys.path`. This lets eval files that are run through the CLI to import
other paths in the project.

- Add a `--terminate-on-failure` option to both PY and JS eval commands,
which terminates the eval runner upon any eval file that itself crashes.
This lets us properly propagate crashes out of the eval runner, which we
can pick up in our own unit tests.