Issue a warning, not an error, if an invalid base experiment is specified (#44)
It's useful to be able to specify base experiments (by name) before they
exist, e.g. in a CI system where you know what they'll be called but
they may run out of order. This changes the SDK to mark them as a
warning (and retry registration) if the base experiment does not exist.