Fix logging issues around metrics. (#45)
- Before, we would crash when a user tried logging something to metrics
in the initial log statement, e.g. `experiment.log(metrics={"foo": 1},
...)`. This is because the SDK would already be trying to log to metrics
internally through `self.initial_data`, and the keys would conflict with
the user trying to log to metrics in their event. We address this by
deep-merging the user's data into `self.internal_data` before logging.

- Users want to be able to override the automatically-populated
"start"/"end" times in a span. We just remove the checks that forbid
this, so they can override any of the metrics.