Introduce braintrust_core package.
This package contains shared dependencies used by both the "braintrust"
package (main SDK) and the "autoevals" package.

Tested by building both the python package and the javascript package.
It seems like we'll need to deploy the package before we can change the
"braintrust" and "autoevals" JS packages to depend on it.