Simplify SDK tracing API (#67)
We make the following improvements/simplifications to the SDK:

- Globally, we maintain a single `currentExperiment` and
`currentLogger`, and a nested `AsyncLocalStorage` `currentSpan`. Now
there is no need for wrapper functions / context managers in experiments
or loggers. But I kept the functions around with a deprecation warning.

- The `traced` decorator in python will auto-log the function's
input/output.

- In javascript, the global `traced` method will now consider `loggers`,
at the cost of making the whole function async.

- Add a `flush` function in place of `close` for experiments and
loggers. Deprecate the `close` method.