JS API: Make all data objects lazily-initialized.
Essentially we accomplish this by logging promises instead of
materialized objects. The logger resolves all the promises before
logging. Intermediate steps share promises, which should just resolve to
the same value once they are awaited, so we're basically using them as a
"lazily-computed value".

We also make the async part of the Logger API optional by introducing a
generic parameter `IsAsyncFlush`. By default, it is false, so all
logging operations will flush and are thus async. But if the user passes
the literal `isAsyncFlush: true`, or the template arg `<true>` (which is
expected to match the true value of `isAsyncFlush`), it will stop
flushing automatically and the type signatures will no longer be wrapped
in promises.