Add support for the feedback API (#89)
This change extends the SDK to support the feedback API, documented at
https://www.braintrustdata.com/docs/guides/logging#user-feedback.

There are two kinds of feedback you can provide:
* The `logFeedback` / `log_feedback` method can be run outside of the
context of a span, and given an `id` + object ids, will add extra
information to a span as a tracked auditable update.
* We also now support creating subspans of older requests, simply by
specifying the object ids and request id. This relies on new
functionality in the API that swaps the `_parent_id` field with the
appropriate `root_span_id` and `span_parents`.