Fix unhandled promise rejections in typescript SDK. (#99)
To accomplish lazy initialization, we were wrapping everything in
promises. The problem with this is that if the promise is not awaited
(more specifically, a rejection handler is not explicitly attached), the
promise's exceptions will get reported as "unhandled" at the end of the
program, as it's clearing out its event loop. This seems to happen even
if the promise is awaited somewhere else and the exception handled (i.e.
in the background logger). I think as long as there is still some
context in which the promise may run without a rejection handler, it
will bubble up at the end.

To work around this, we add one more step of indirection, now similar to
what we do in the python SDK. The LazyValue wrapper creates the promise
only when asked for, and the user is expected to await it immediately.
Any delayed computations that are passed around as arguments or used in
other lazy computations should be wrapped in LazyValue rather than as a
bare promise.

Fixes BRA-890