Update "output" field on datasets to "expected" (#100)
Updates the SDK to use `expected` instead of `output` for dataset
records.

The object fetcher and logger now attempt to use new API endpoints
(`logs3/` and `object3/`) that use `expected` instead of `output` for
datasets. If the request errors, we fall back to the old endpoints,
patching the data to use `output` accordingly.