Robustness and debuggability improvements to background loggers (#125)
- Add retry logic to the LazyValue resolution when logging.

- Introduce a `sync_flush` parameter to the loggers. If false (the
current behavior), any lazy value resolution and logging is run
effectively in the background, so that any errors are just printed out
after all retries have completed. If true, the background thread is
inactive. We still print out the errors when retrying, but after the
retries have been exhausted, we throw an exception. This should in
theory give users fine-grained control over which/when statements are
sent to the network, and gives them the ability to implement their own
control flow.

- While we're at it, make the other logger arguments controllable via
environment variable.

- While implementing the same functionality in the JS logger, I took a
pass at trying to more closely mirror the logic between python and JS.