Log objects eagerly
It was pointed out that objects logged to braintrust may be mutated
after logging, and braintrust would end up logging the mutated version.
We fix this by round-tripping the object through the JSON serializer,
which doubles as a synchronous check for serializability (that we were
already doing in python).

We also adjust the `@traced` decorator in python to log the input before
running the function, so that it's immediately visible.