Span logging: use parent_id to set parent info everywhere.
Mixing parent_id with explicit root_span_id/span_id setting leads to
messy outcomes, because if someone specifies a `parent_id` field, we can
no longer rely on the autogenerated `root_span_id` field that we set in
the SDK, but we'll still propagate that down to sub-spans, leading to a
bad outcome.

The correct "intermediate" fix is to stop relying on explicit parent
propagation as soon as the span is initialized with a parent_id.