Fix bug in JS LazyValue (#140)
Since the `.get` method was async, it was possible that multiple callers
could invoke it concurrently. Also it was awaiting the result of the
callable in the same statement as setting the callable to computed,
which allows multiple callers to generate and await the callable
concurrently.

The observed result is that the LazyValue function could be invoked
multiple times, and we might end up registering multiple experiments.
The new implementation should guarantee we invoke the callable at most
once.