More sophisticated batching for handling `PARENT_ID_FIELD` references
Currently, the logic in the various loggers (both in the SDK and
internally) assume that once you run `merge_row_batch` to consolidate log
entries with the same ID, you can log all the rows in parallel.

Since we introduced the `PARENT_ID_FIELD` attribute to allow creating a
span underneath another row, where the reference is resolved on the
backend, we now have a dependency between parent and child log entries,
where the parent must be logged before the referencing child.

We address this in two places: in `merge_row_batch`, we bucket all of
the merged rows into connected components based on `PARENT_ID_FIELD`
relationships, and then topologically sort them such that parents come
before children in each bucket. Elements within each bucket must be
processed in that order, while elements in different buckets can be
logged/processed in parallel.

We also rework the `batchRecords` logic to operate over this bucketed
output of `merge_row_batch`. This is slightly tricky because it is not
always possible to flush a full ordered bucket at once to the backend
due to payload size constraints. So if an ordered bucket does not fit
entirely into one batch, we must defer publishing it until we've
published the earlier elements in the bucket. So now the batching logic
returns a triple-nested `List[List[List[str]]]`, where the outermost
`List[List[str]]` batch sets must be published in sequence, while each
`List[str]` batch within a batch set can be published in parallel, and
each `List[str]` batch individually is dependency-ordered.