Fix handling of parent_id in `SpanImpl` (#143)
Once you initialize a span with `PARENT_ID_FIELD`, we can no longer
trust the autogenerated `root_span_id`, since it will be resolved on the
backend. But we noticed a bug that if you create a sub-span from the
span, it would propagate that incorrect `root_span_id` down and the
subspan would have the wrong parent.

We fix this by only populating `span_id` and `root_span_id` if you have
directly supplied parent info, otherwise only populating
`PARENT_ID_FIELD`. Then when creating a sub-span, we basically propagate
either the parent_id or the provided span ids.