wrap the iterator only in the OAI stream interface (#147)
The stream object in the openAI api has some other functionality other
than just being an `AsyncIterator` (e.g. `toReadableStream`, `tee`,
etc.). So just wrap the iterator itself and return the base open ai
stream object

Not sure if this is what you were looking for but it seems to work.