Remove strictness requirements on API typespecs.
These typespecs are part of the OpenAPI spec that we publish, and while
it's nice to have automated checks in the backend API server to ensure
the user doesn't specify unnecessary properties, this strictness makes
it hard to add properties to various operations without forcing folks to
upgrade their backend.

So we remove the strictness requirements now. We can consider adding the
checks back later, perhaps in a more ad-hoc way, if they help users.