Allow exporting parent object + row id for starting a subspan. (#159)
Originally we introduced the `parent_id` field as a way to start a
subspan on the same toplevel object using the specified id as the span
parent. But it's more convenient if folks can export the whole
`(object_type, object_id, row_id)` tuple and start a subspan from that.
To accomplish this, we create a serialization format for that tuple,
which can be obtained via `[Experiment|Logger|Span].export_as_parent`.
Then we introduce a new argument `parent` for starting a span from an
exported parent string. We deprecate the `parent_id` field in the SDK,
because it's essentially redundant with `parent` and less functionaly.

Along the way, we make some cleanups to the SDK implementation,
primarily around handling and passing around parent information. Since
we made some important correctness fixes to the `parent_id` field on the
backend in API version `0.0.35`, using this new SDK will require people
to upgrade their backend to at least this version.