Use single global BackgroundLogger instance everywhere. (#179)
Originally we would spawn a separate BackgroundLogger for each object we
created. This is not particularly useful anymore, because the background
logger already batches and parallelizes requests. It also leads to
problems with flushing, because you have to call flush on every logger
that you have logged to (including the global logger if you are doing
distributed tracing).

We change everything to use just the single global logger and add a
global flush function for good measure. Fix a few bugs along the way.