Adjustments for new `update` semantics.
Internally, we are adjusting the `update` param in experiment
registration to create the experiment if it doesn't exist and return it
unmodified if it does exist. We clarify these semantics in the docs, and
go back to running "register experiment" with the update parameter.

We are also changing the behavior of the REST API to default to
returning the existing experiment if one already exists, unless you
specify `ensure_new=true`.