Change `span.export()` to log more information. (#196)
It was pointed out that the current approach to distributed tracing is
prone to race conditions between the parent span logger and the subspan
logger: if the subspan hits the backend first, it will fail to nest
underneath the parent span because the backend tries to resolve the span
pointer information by doing a lookup for the parent span id.

To fix this, we avoid the whole backend-pointer-resolution logic by
explicitly providing the span id and root span id when logging spans
(this was actually what we did originally). We also modify the slug
produced by `span.export()` to include the span id and root span id, and
encode it as a more-opaque base64 string to reduce string length.