Robustify the `patch` endpoint for groups and roles. (#229)
While implementing the RBAC UI internally, we discovered that the
current PATCH API is prone to race-conditions, because adding/removing
members of a group or role requires reading the full list and then
replacing it with an updated list. We replace the replace-list params
with separate add/remove-from-list params to fix this.

If people truly want to replace the lists, they can use a `PUT` request.