Improve IPython repr for EvalResultWithSummary (#233)
Prior to this change, when you ran `Eval()` in a notebook, it'd spit out
a full summary along with all the results. This just cleans up the
`_repr_pretty_` implementation to print a tighter object.

Note that we print (to the console) the summary itself, so if we printed
the summary itself, it'd show up twice.