Logger: don't block span export on resolving id
Currently, logging is fully non-blocking (in that it doesn't need to hit
the network on the critical path), except for the `[Logger/Span].export`
function, which needs to resolve the object ID to serialize for the
distributed tracer.

In order to work around this and make the export function non-blocking,
we pass the arguments to the logger's `computeMetadata` registration
function into the `Logger` and include them in the serialized span
identifier if the object ID has not yet been resolved. Then when
deserializing the parent span slug, we can obtain the object ID either
from the serialized ID or by reconstructing the `computeMetadata` lambda
from the serialized args.