Add a prompt fallback (#281)
This is likely not the final solution... but allow the messages array to
have a fallback for arbitrary roles (e.g. "model" from google), so that
in the playground, we permit non-openai-compatible sessions.

`prompt.build()` still needs to produce something that typechecks with
the OpenAI API, so still expose `OpenAIMessages` and cast
`prompt.build()` to that. Note that with this approach, you can
(incorrectly) cast a Google prompt to OpenAI.