Only mark braintrust/autoevals as external (#284)
I don't know why (and git doesn't seem to tell more) why we did a
catch-all filter here, but it breaks some complex bundling scenarios
(exposed in a test app we're working on). The intent was to exclude
braintrust-flavored packages, so I made the regex more specific and
manually confirmed.