Replace all usages of `strictObject` with `object`.
We've decided that by default, our zod object schemas should be in
`strip` mode rather than `strict` mode, because there are many usage
sites which only want to use the schema for validation, which may not be
always up-to-date with the latest schema definition (e.g. the API
server, sdk libraries). So `strip` is safer for validation.

For users who do want to check no extra fields, we introduce a
`parseNoStrip` utility function which parses the object and checks that
no fields were removed from the final object.

TODO before merging:

- Documentation about our schema definition rationale
- Unit tests for util functions