Native support for streams (#288)
This turned into a fairly large change, which adds support for a variety
of different streaming operations. Here's a summary of stuff to look at:

Main:

* Typespecs for manipulating functions/invoking them (these were moved
into the repo from the main repo)
* A new `@braintrust/vercel-ai-sdk` package which implements our
adapter.
* functions/invoke.ts calls the invoke endpoint
* Propagate proxyUrl as state we derive from logging in
* Use `waitUntil` in Vercel to make turning `asyncFlush: true` safe. We
may want to switch the default now...
* Add `BraintrustStream` and some basic tools to parse/transform it.
This is the streaming type returned via `invoke`
* Support logging streams

Misc:

* Upgraded bundling in sdk to use `tsup` so we don't need to deal with
inline esbuild flags
* Made the `zod` dependency external, so that you can BYO zod and use
type inference with the SDK (first used in the new `invoke` function)
* Hand-coded the `braintrust` package's version as 0.0.141, so that the
`vercel-ai-sdk` package, when published, uses it. We have some follow up
work to do to fix these versions.