Fix bug in `computeLoggerMetadata` (#339)
This function is invoked when lazily initializing the logger object's
metadata, either when initializing a logger or when logging to a parent
span that originates from a logger. In the former case, we were already
logging in with the user's provided params, but in the second case, we
still need to log in.

This matches the implementation in the python `logger.py`, so it was
just an accidental omission.