Revise span components slug again (#359)
We need to support adding arbitrary properties. So basically now the
serialized object representation consists of (1) a few required fields
(2) a set of fields we can serialize as compressed UUIDs (3) a JSON
object for any remaining fields. Hopefully this is extensible enough to
avoid further revisions.

Tested out that the slug size is roughly the same as before.