Add explicit "disabled" state for async scoring.
Sometimes we want to log a row (or a trace) where we explicitly exclude
it from async scoring. This is not easy to do right now, because you'd
have to pass `SKIP_ASYNC_SCORING: true` every time you do any
incremental logging. But now we can override the state to disabled once
at the beginning and leave it.