Decouple root span id from span id (#389)
We still set root_span_id to the root span's span_id for now, but we are
gearing up for being able to decouple them without breaking stuff. The
new `is_root` field for events is basically equivalent to
`len(span_parents) == 0` but is added for quicker access in some cases.