Add support for uploading arbitrary file attachments.
We introduce a new class, `LazyAttachmentReference`, which represents
the upload parameters for a file on disk or in-memory Blob/Buffer and
can be constructed synchronously. The caller can place
`LazyAttachmentReference` objects anywhere in the event to be logged.

```ts
const myFile: Uint8Array = ...;
currentSpan().log({
  input: {
    question: "What is this?",
    context: {
      attachment: new LazyAttachmentUpload(myFile, "bridge.jpg", "image/jpeg"),
    },
  },
});
```

At span upload time, the SDK replaces lazy attachments with a reference
to the object store. The above call is equivalent to:

```ts
currentSpan().log({
  input: {
    question: "What is this?",
    context: {
      attachment: {
        filename: "bridge.jpg",
        content_type: "image/jpeg",
        path: "attachments/19/1935ec5b-51fd-4267-a2af-db38ee37f46e",
      },
    },
  },
});
```

(Users who desire more control may handle the upload on their own and
pass the resulting reference.)

At the same time, the upload is also queued. This implementation ensures
that log insertion can happen in parallel with attachment uploads.