Add support for uploading file & blob attachments in JS. (#404)
We introduce a new class, `Attachment`, which represents the upload
parameters for a file on disk or in-memory Blob/Buffer and can be
constructed and JSON serialized synchronously. The caller can place
`Attachment` objects anywhere in the event to be logged.

```ts
const myFile: Uint8Array = ...;
currentSpan().log({
  input: {
    question: "What is this?",
    context: {
      attachment: new Attachment({
        data: myFile,
        filename: "bridge.jpg",
        contentType: "image/jpeg",
      }),
    },
  },
});
```

Prior to event upload, the SDK serializes `Attachment` objects which
replaces them with a reference to the object store. The above call is
equivalent to:

```ts
currentSpan().log({
  input: {
    question: "What is this?",
    context: {
      attachment: {
        type: "braintrust_attachment",
        filename: "bridge.jpg",
        content_type: "image/jpeg",
        key: "1935ec5b-51fd-4267-a2af-db38ee37f46e",
      },
    },
  },
});
```

This serialization is synchronous, so the presence of attachments
doesn't affect upload performance.

Users who desire more control can upload the file ahead of time by
retaining a reference to the `Attachment` object.