Fix logging assumptions
We used to try to autodetect or inititialize a logger. This was good DX,
but in the event the user is running their code in an Eval it did
not emit the spans to the eval session.

Now, we'll continue to accept an optional logger, but we'll use `startSpan`
which will check for an experiment session or use a project logger or
fallback to a noop span. In the case, that nothing is configure, we
will warn the user that they'll want to configure logging.

**Note**: regressed the attempt to add tags to the root span. If running
in an experiment, we may not have the root span available to add tags.
Instead, the tags will go to the span's metadata.