Make `span.export()` nonthrowing when braintrust is down. (#548)
This was actually already true in the python logger, but in typescript
we were marking the promise as `hasComputed` before it had resolved
successfully. So when running `span.export()`, if the project id has
lazily resolved to an error, we would await the promise and throw.

Now we change the contract so that we mark the LazyValue as
`isSucceeded` when the promise resolves successfully. Before that,
`span.export` will embed the object constructor arguments.