Print but do not fail when scorers fail (#565)
We historically fail the whole eval case when a scorer fails. But now
that we can report errors more granularly (per scorer), we don't really
need to do this.

We still print the error to the terminal, but now let just the
individual scorer fail.

In the UI, it looks like this:

![CleanShot 2025-03-01 at 10 09
50](https://github.com/user-attachments/assets/fb04ef8f-2a57-43ab-aa63-dd83f2ddc1f8)