Add nox tests.
Nox is a testing tool that we are using to test against different
versions of the anthropic library. Crucially it will always test against
the latest version, so we'll catch any newly released packages that we
don't play nicely with.

Nox can handle different versions of Python also, but we are doing that
one level higher so I'm ignoring that for now.