anthropic.ts (#626)
Typescript tracing for the anthropic SDK. It includes all the obvious
tracing code + tests, which should be reviewed.

It also includes a bunch of changes to facilitate testing.

- Added `_exportsForTestingOnly.simulateLoginForTests` which allows us
to skip the http requests for the server for testing. This changes
`logger.ts` so worth looking at.
- added `TestBackgroundLogger` to intercept logs and save to memory. It
also includes some hooks on `_exportsForTestingOnly` to use it.
- Added a crude way of running our tests with different versions of the
`anthropic-sdk` library. This is all in the `Makefile`. This works, but
is there a better way?
- Added some example files which are nice for customers and to spot
check the data we create