Remove fallback to legacy `/logs` endpoint when logging
Data planes have supported the `/logs3` endpoint for long-enough that we
don't expect users of new SDKs to rely on this fallback. The fallback
essentially doubles the number of retries against the data plane from 3
to 6, so this will also bring that down to 3. We can consider re-bumping
that if necessary.