Pin Python Build Dependencies (#698)
This PR makes a few changes:

- Pins all the of the critical build dependencies. This will help ensure
that a new version of build, twine or setuptools doesn't have unexpected
behaviour that changes our app.
- splits the dependencies by function:
- `requirements-dev.txt` is the dependencies developers will want to
install on their machine. These are loosely managed. `pylint`, `black`,
etc.
- `requirements-build.txt` is the dependencies we need to build and ship
the SDK. Everything in here is pinned and strictly managed. `build`,
`setuptools`, etc.
- `setup.py` is for actual dependencies of our library. I'm happy the
dev stuff is gone from here because it's easier to read and understand.
  - Installs `pip` in a fixed way.
  - Updates github actions to install things the right way.