Use typespecs generated from openapi
This PR is part of an effort to move ownership of the typespecs in
`@braintrust/core/typespecs` into our internal repo, and expose them to
our public repos through an openapi spec. Generally this makes sense
because most typespecs reflect the APIs provided by our internal
services, so they ought to be owned by the internal services.

Aside from the new auto-generated typespecs, most of the changes to
existing code are just modifying imports to point to the auto-generated
typespecs. And most changes to the typespecs code are just explicitly
defining names for schemas that we use in either the JS or PY code, so
that they get picked up by the openAPI generator.

The main trickiness involves handling zod constructs which transform
data from an input type to an output type, e.g. `.default`,
`.transform`, `.preprocess`, etc. Such schemas have a different "input"
type than their "output" type, and our current openapi spec generator
seems to generate based on input types. So in certain cases, we tack on
some properties to the openapi spec to instruct our generator to
post-process the spec in certain ways, so that it aligns with the output
type. In the future, we can probably refactor our zod -> openapi
pipeline to use an output-type-sensitive generator.