Move `@braintrust/core` into `braintrust` package
Currently every package that depends on `@braintrust/core` also depends
on `braintrust`. So there's no real value in having a separate package.
This PR moves the contents of `@braintrust/core` into a subdirectory of
the `braintrust` package, that can be imported as `braintrust/util`.
Removing the extra package should make deployment overall easier.