enable mixed otel & braintrust tracing
This PR enables mixed otel & braintrust tracing.

The main use cases are
    (a) tracing otel in evals
    (b) a building block towards enabling distributed tracing between braintrust & otel managed processes.

It works by setting the env var BRAINTRUST_OTEL_COMPAT=true. If not enabled, nothing should change.

It works by sharing the "is there an active span that should be my parent?" decision with otel. In
technical terms, we write active braintrust spans (wrapped in a fake otel span)
to the otel contextvars context and always read from the otel span.

It piggybacks on #938 which allows otel compatible IDs.