first attempt at enabling reasoning
this worked to an extent, but i keep finding a few cases where very
simple single shot generation we do not provide the reaosning response.

this is because we are doing both the generateText and doGenerate
(middleware) which i would argue is very confusing to the user.

therefore, i'm planning on simplifying our integration grately.
committing this to leave a trace, but going to move on to a more
aggressive change