[JS-SDK] Smoke tests for the js-sdk in different environments (#1069)
### Background

A while ago the ESM version of the build broke for users of Otel. No one
knew because there are no tests around the failure.

### Changes

This PR adds integration tests to the build pipeline that are currently
allowed to fail. Once the ESM issue has been resolved the
continue_on_error setting can be removed and the tests will run on the
pipeline to verify the changes.

There are three tests:
- a simple span test that will run in CJS and then ESM
- an otel test that will run in CJS and then ESM
- the simple span test run in a deno environment to capture ESM native
failures