chore: Explicit exports (#1168)
We have multiple public interfaces:

- `/` which is the main point of entry
- `/dev` for the dev / remote evals
- `/browser` which is a more minimal sdk that can be used in more
restricted environments.

When we worked on 1.x, we found a few cases where we neglected to export
new wrappers in the right file. Part of this is that we have complicated
export process.

This makes that a little more explicit. Ideally, we wouldn't have the
exports- files and the entrypoints would have the literal exports that
we want to support in those environments.

For now, this moves us away from `export *` and now to explicit exports.
This also adds new rules to prevent us from regressing and doing `export
*`.

---------

Co-authored-by: Ankur Goyal <ankrgyl@gmail.com>